Fix Android versions below 3.1. Joystick code that was added (OnGenericMotionListener) requires Android 3.1. Joystick won't work below 3.1 but everything else will now work.
