+++ android/allegro_activity/src/AllegroSurface.java
@@ -7,11 +7,8 @@ import android.view.Display;
 import android.view.SurfaceHolder;
 import android.view.SurfaceView;
 import android.view.View;
-import android.view.View.OnGenericMotionListener;
-import android.view.MotionEvent;
-import android.view.InputDevice;
 
-class AllegroSurface extends SurfaceView implements SurfaceHolder.Callback, OnGenericMotionListener
+class AllegroSurface extends SurfaceView implements SurfaceHolder.Callback
 {
    /** native functions we call */
    public native void nativeOnCreate();
@@ -21,6 +18,7 @@ class AllegroSurface extends SurfaceView implements SurfaceHolder.Callback, OnGe
    public native void nativeOnJoystickButton(int index, int button, boolean down);
 
    private AllegroActivity activity;
+   private AllegroJoystick joystick_listener;
 
    /** functions that native code calls */
 
@@ -103,7 +101,11 @@ class AllegroSurface extends SurfaceView implements SurfaceHolder.Callback, OnGe
       requestFocus();
       setOnKeyListener(key_listener);
       setOnTouchListener(touch_listener);
-      setOnGenericMotionListener(this);
+
+      if (android.os.Build.VERSION.SDK_INT >= 12) {
+         joystick_listener = new AllegroJoystick(activity, this);
+         setOnGenericMotionListener(joystick_listener);
+      }
    }
 
    @Override
@@ -150,75 +152,6 @@ class AllegroSurface extends SurfaceView implements SurfaceHolder.Callback, OnGe
    void setCaptureVolumeKeys(boolean onoff)
    {
       key_listener.setCaptureVolumeKeys(onoff);
-   }
-
-   private float axis0_x = 0.0f;
-   private float axis0_y = 0.0f;
-   private float axis0_hat_x = 0.0f;
-   private float axis0_hat_y = 0.0f;
-   private float axis1_x = 0.0f;
-   private float axis1_y = 0.0f;
-
-   private void handleHat(int index1, float old, float cur, int button1, int button2) {
-      if (old == cur)
-         return;
-
-      if (old == 0) {
-         if (cur < 0)
-            nativeOnJoystickButton(index1, button1, true);
-         else
-            nativeOnJoystickButton(index1, button2, true);
-      }
-      else if (old < 0) {
-         nativeOnJoystickButton(index1, button1, false);
-         if (cur > 0) {
-            nativeOnJoystickButton(index1, button2, true);
-         }
-      }
-      else if (old > 0) {
-         nativeOnJoystickButton(index1, button2, false);
-         if (cur < 0) {
-            nativeOnJoystickButton(index1, button1, true);
-         }
-      }
-   }
-
-   @Override
-   public boolean onGenericMotion(View v, MotionEvent event) {
-      if (activity.joystickActive == false) {
-         return false;
-      }
-
-      int id = event.getDeviceId();
-      int index = activity.indexOfJoystick(id);
-      if (index >= 0) {
-         float ax = event.getAxisValue(MotionEvent.AXIS_X, 0);
-         float ay = event.getAxisValue(MotionEvent.AXIS_Y, 0);
-         float ahx = event.getAxisValue(MotionEvent.AXIS_HAT_X, 0);
-         float ahy = event.getAxisValue(MotionEvent.AXIS_HAT_Y, 0);
-         float az = event.getAxisValue(MotionEvent.AXIS_Z, 0);
-         float arz = event.getAxisValue(MotionEvent.AXIS_RZ, 0);
-         if (ax != axis0_x || ay != axis0_y) {
-            nativeOnJoystickAxis(index, 0, 0, ax);
-            nativeOnJoystickAxis(index, 0, 1, ay);
-            axis0_x = ax;
-            axis0_y = ay;
-         }
-         else if (ahx != axis0_hat_x || ahy != axis0_hat_y) {
-            handleHat(index, axis0_hat_x, ahx, AllegroActivity.JS_DPAD_L, AllegroActivity.JS_DPAD_R);
-            handleHat(index, axis0_hat_y, ahy, AllegroActivity.JS_DPAD_U, AllegroActivity.JS_DPAD_D);
-            axis0_hat_x = ahx;
-            axis0_hat_y = ahy;
-         }
-         if (az != axis1_x || arz != axis1_y) {
-            nativeOnJoystickAxis(index, 1, 0, az);
-            nativeOnJoystickAxis(index, 1, 1, arz);
-            axis1_x = az;
-            axis1_y = arz;
-         }
-         return true;
-      }
-      return false;
    }
 }
